refactor extension app components
